// src/subapps/projects/config.d.ts
export declare const fusionConfig: {
  configOrg: string;
  configProject: string;
  personalOrgPrefix: string;
  workflowStepType: string;
  activityWorkflowLink: string;
  templateType: string;
  projectMetadataType: string;
  fusionProjectTypes: string[];
  codeType: string[];
  noteType: string[];
  agentOrgType: string[];
  agentPersonType: string[];
  datamodelsActivityId: string;
  defaultAPIMappings: {
    prefix: string;
    namespace: string;
  }[];
};
export default fusionConfig;